smartcompute_hrm_proto/python/requirements.txt
cryptography>=39.0.0
PyYAML>=6.0
requests>=2.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
